Music stretching
Music can now be captured cleanly, for the most part. There are some issues with tick advance still, need to look into it more.

+++ Source/Cordyceps2/TimeControl.cs
@@ -15,7 +15,7 @@ public static class TimeControl
     // default tickrate is 30 due to being in the depths, and TimeControl is slowing the tickrate down to 20, then this
     // will be 20 / 30.
     // TODO: Reset when RainWorldGame is no longer main loop
-    public static double ArtificialTimeFactor = 1.0;
+    public static float ArtificialTimeFactor = 1.0f;
     
     public static bool TickrateCapOn;
     public static bool TickPauseOn;
@@ -61,12 +61,12 @@ public static class TimeControl
 
                 if (!CanAffectTickrate())
                 {
-                    ArtificialTimeFactor = 1.0;
+                    ArtificialTimeFactor = 1.0f;
                     return;
                 }
                 
                 var targetTickrate = TickPauseOn ? 0 : Math.Min(DesiredTickrate, game.framesPerSecond);
-                ArtificialTimeFactor = UnmodifiedTickrate == 0 ? 0.0 : targetTickrate / (double)UnmodifiedTickrate;
+                ArtificialTimeFactor = UnmodifiedTickrate == 0 ? 0.0f : targetTickrate / (float)UnmodifiedTickrate;
                 game.framesPerSecond = targetTickrate;
             }
             catch (Exception e)
